# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aurora Hotels — Hotel Management System API",
        "description": "Versioned REST API for authentication, bookings, guests, rooms, invoices, payments, and reports (FR-API-001). All responses use consistent API Resource transformers; all errors share a single {\"error\": {code, message, errors?}} envelope (FR-API-003).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Default API server"
        }
    ],
    "paths": {
...
        "/room-types": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "List room types for a branch",
                "operationId": "81632eb38c70d50132dc6650f0d4f3e2",
                "parameters": [
                    {
                        "name": "branch_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room types",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/RoomType"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/room-types/{roomType}/availability": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Check room availability for a date range",
                "operationId": "f4caca0a6bb2d86b2f4b1f0ba0be31f0",
                "parameters": [
                    {
                        "name": "roomType",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "arrival_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "departure_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Availability",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "room_type_id": {
                                                    "type": "integer"
                                                },
                                                "available_rooms": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ErrorEnvelope": {
                "properties": {
                    "error": {
                        "properties": {
                            "code": {
                                "type": "string",
                                "example": "validation_error"
                            },
                            "message": {
                                "type": "string",
                                "example": "The given data was invalid."
                            },
                            "errors": {
                                "type": "object",
                                "nullable": true,
                                "additionalProperties": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "RoomType": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "branch_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Deluxe"
                    },
                    "slug": {
                        "type": "string",
                        "example": "deluxe"
                    },
                    "base_capacity_adults": {
                        "type": "integer",
                        "example": 2
                    },
                    "base_capacity_children": {
                        "type": "integer",
                        "example": 1
                    },
                    "base_rate_cents": {
                        "type": "integer",
                        "example": 15000
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctum personal access token. Send as `Authorization: Bearer {token}`. Tokens are scoped by ability (FR-API-004) — see the abilities returned at login.",
                "bearerFormat": "Opaque token issued by POST /auth/login",
                "scheme": "bearer"
            }
        }
    },
...
}